# Conflicts:
#	cmp-shared/src/commonMain/kotlin/org/mifospay/shared/navigation/MifosNavHost.kt
#	core/designsystem/src/commonMain/kotlin/org/mifospay/core/designsystem/icon/MifosIcons.kt
#	feature/mpay-qr-scan/src/commonMain/kotlin/org/mifospay/feature/mpay/qr/scan/di/MpayQrScanModule.kt
#	feature/mpay-qr-scan/src/commonMain/kotlin/org/mifospay/feature/mpay/qr/scan/navigation/MpayQrScanNavigation.kt
#	feature/qr/src/commonMain/kotlin/org/mifospay/feature/qr/ScanQrCodeScreen.kt
#	feature/qr/src/commonMain/kotlin/org/mifospay/feature/qr/ScanQrViewModel.kt
#	feature/send-money/src/androidMain/kotlin/org/mifospay/feature/send/money/QrScanner.android.kt
#	feature/send-money/src/commonMain/composeResources/values/strings.xml
#	feature/send-money/src/commonMain/kotlin/org/mifospay/feature/send/money/SendMoneyScreen.kt
#	feature/send-money/src/commonMain/kotlin/org/mifospay/feature/send/money/SendMoneyViewModel.kt
#	feature/send-money/src/commonMain/kotlin/org/mifospay/feature/send/money/navigation/SendNavigation.kt

+++ core/model/src/commonMain/kotlin/org/mifospay/core/model/savingsaccount/Summary.kt
@@ -17,10 +17,10 @@ data class Summary(
     val currency: Currency,
     val totalDeposits: Double = 0.0,
     val totalWithdrawals: Double = 0.0,
-    val totalInterestPosted: Long = 0,
+    val totalInterestPosted: Double = 0.0,
     val accountBalance: Double = 0.0,
-    val totalOverdraftInterestDerived: Long = 0,
-    val interestNotPosted: Long = 0,
+    val totalOverdraftInterestDerived: Double = 0.0,
+    val interestNotPosted: Double = 0.0,
     val availableBalance: Double = 0.0,
 )
